+++ hooks/checkstyle_jar.py
@@ -15,17 +15,20 @@ import argparse
 import pathlib
 import sys
 from pathlib import Path
-from subprocess import run
+from subprocess import run, CompletedProcess
 from urllib import request, parse
+from urllib.parse import ParseResult
+from http.client import HTTPMessage
+from typing import List, Optional, Sequence
 
-RC_OK = 0
-RC_ERROR = 1
-RC_FAKE_ERROR_CODE = 10
-RC_DOWNLOAD_ERROR = 20
-RC_NO_JAR = 150
+RC_OK: int = 0
+RC_ERROR: int = 1
+RC_FAKE_ERROR_CODE: int = 10
+RC_DOWNLOAD_ERROR: int = 20
+RC_NO_JAR: int = 150
 
 
-def main(argv = None):
+def main(argv: Optional[Sequence[str]] = None) -> int:
     parser = argparse.ArgumentParser()
     command_group = parser.add_mutually_exclusive_group()
     command_group.add_argument(
@@ -42,19 +45,19 @@ def main(argv = None):
         '--config', dest = 'config', nargs = 1, action = 'store', metavar = "PATH", default = '/google_checks.xml',
         help = 'Path to checkstyle config to use. Use "/google_checks.xml" or "/sun_checks.xml" to use built-in styles.')
     parser.add_argument('files', nargs = '*', help = 'Files to check.')
-    args = parser.parse_args(argv)
+    args: argparse.Namespace = parser.parse_args(argv)
 
     if args.jar_url:
-        args.jar_url = args.jar_url[0]
+        args.jar_url: str = args.jar_url[0]
 
     # https://github.com/checkstyle/checkstyle/releases/
     if args.jar:
-        args.jar = Path(args.jar[0]).expanduser()
+        args.jar: Path = Path(args.jar[0]).expanduser()
 
     if not args.jar_url and not args.jar:
-        args.jar_url = 'https://github.com/checkstyle/checkstyle/releases/download/checkstyle-10.4/checkstyle-10.4-all.jar'
+        args.jar_url: str = 'https://github.com/checkstyle/checkstyle/releases/download/checkstyle-10.4/checkstyle-10.4-all.jar'
 
-    args.cache = Path(args.cache[0]).expanduser()
+    args.cache: Path = Path(args.cache[0]).expanduser()
     if not args.cache.is_dir():
         print(f'The --cache must point to writable directory: {args.cache}')
         return RC_ERROR
@@ -63,21 +66,23 @@ def main(argv = None):
         return RC_OK
 
     if args.jar_url:
-        parsed_url = parse.urlparse(args.jar_url)
-        downloaded_jar_filename = pathlib.Path(parsed_url.path).name
-        downloaded_jar_path = Path(args.cache).expanduser() / downloaded_jar_filename
-        downloaded_tmp_path = Path(args.cache).expanduser() / f'{downloaded_jar_filename}.tmp'
+        parsed_url: ParseResult = parse.urlparse(args.jar_url)
+        downloaded_jar_filename: str = pathlib.Path(parsed_url.path).name
+        downloaded_jar_path: Path = Path(args.cache).expanduser() / downloaded_jar_filename
+        downloaded_tmp_path: Path = Path(args.cache).expanduser() / f'{downloaded_jar_filename}.tmp'
         if not downloaded_jar_path.exists():
             print(f'Downloading {downloaded_jar_filename} to {args.cache}...')
+            path: str
+            http: HTTPMessage
             path, http = request.urlretrieve(args.jar_url, downloaded_tmp_path)
-            tmp_path = Path(path)
+            tmp_path: Path = Path(path)
             if not tmp_path.exists():
                 print(f'Failed to download JAR file: {args.jar_url}')
                 return RC_DOWNLOAD_ERROR
             tmp_path.rename(downloaded_jar_path)
 
         if downloaded_jar_path.exists():
-            args.jar = downloaded_jar_path
+            args.jar: Path = downloaded_jar_path
 
     if not args.jar.exists():
         print(f'Checkstyle JAR file not found: {args.jar}')
@@ -87,13 +92,13 @@ def main(argv = None):
         return RC_NO_JAR
 
     # https://checkstyle.sourceforge.io/cmdline.html#Download_and_Run
-    cmd = ['java', '-jar', args.jar, '-c', args.config] + args.files
+    cmd: List[str] = ['java', '-jar', str(args.jar), '-c', args.config] + args.files
 
-    completed = run(cmd, capture_output = True)
-    return_code = completed.returncode
+    completed: CompletedProcess = run(cmd, capture_output = True)
+    return_code: int = completed.returncode
 
-    stdout = list(filter(lambda item: item != '', completed.stdout.decode().split('\n')))
-    stderr = list(filter(lambda item: item != '', completed.stderr.decode().split('\n')))
+    stdout: List[str] = list(filter(lambda item: item != '', completed.stdout.decode().split('\n')))
+    stderr: List[str] = list(filter(lambda item: item != '', completed.stderr.decode().split('\n')))
     if stderr:
         print("\n".join(stderr))