+++ CHANGES.md
@@ -6,6 +6,9 @@ Handy Git hooks to integrate with [pre-commit](http://pre-commit.com/) framework
 
 # Changelog #
 
+* dev
+  * Added type hints to hooks code.
+
 * v1.3.1 (2022-11-07)
   * Updated docs to include `exclude_types` example (i.f. for binaries).
   * Bumped default version of Checkstyle to 10.4